requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
asyncpg
psycopg2-binary
passlib
python-jose
python-dotenv
